+++ app/Repositories/RoleRepository.php
@@ -2,7 +2,9 @@
 
 namespace App\Repositories;
 
+use App\Exceptions\NotFoundException;
 use App\Interfaces\RoleInterface;
+use App\Models\Permission;
 use App\Models\Role;
 use App\Models\User;
 
@@ -15,7 +17,7 @@ class RoleRepository implements RoleInterface
 
     public function getRoleById($id)
     {
-        return Role::with('permissions')->findOrFail($id);
+        return Role::with('permissions')->find($id);
     }
 
     public function createRole(array $data)
@@ -25,7 +27,10 @@ class RoleRepository implements RoleInterface
 
     public function updateRole($id, array $data)
     {
-        $role = Role::findOrFail($id);
+        $role = Role::find($id);
+        if (! $role) {
+            return null;
+        }
         $role->update($data);
 
         return $role;
@@ -33,14 +38,25 @@ class RoleRepository implements RoleInterface
 
     public function deleteRole($id)
     {
-        $role = Role::findOrFail($id);
+        $role = Role::find($id);
+        if (! $role) {
+            return false;
+        }
 
         return $role->delete();
     }
 
     public function assignPermissionToRole($roleId, $permissionName)
     {
-        $role = Role::findOrFail($roleId);
+        $role = Role::find($roleId);
+
+        if (! $role) {
+            return null;
+        }
+        if (! Permission::where('name', $permissionName)->exists()) {
+            throw new NotFoundException('Permission not found');
+        }
+
         $role->givePermissionTo($permissionName);
 
         return $role;
@@ -48,7 +64,10 @@ class RoleRepository implements RoleInterface
 
     public function removePermissionFromRole($roleId, $permissionIds)
     {
-        $role = Role::findOrFail($roleId);
+        $role = Role::find($roleId);
+        if (! $role) {
+            return null;
+        }
         $role->revokePermissionTo($permissionIds);
 
         return $role;
@@ -56,7 +75,15 @@ class RoleRepository implements RoleInterface
 
     public function assignRoleToUser($data): User
     {
-        $user = User::findOrFail($data['user_id']);
+        $user = User::find($data['user_id']);
+
+        if (! $user) {
+            throw new NotFoundException('User not found');
+        }
+
+        if (! Role::where('name', $data['role_name'])->exists()) {
+            throw new NotFoundException('Role not found');
+        }
 
         $user->assignRole($data['role_name']);
 
@@ -67,6 +94,10 @@ class RoleRepository implements RoleInterface
     {
         $user = User::findOrFail($data['user_id']);
 
+        if (! Role::where('name', $data['role_name'])->exists()) {
+            throw new NotFoundException('Role not found');
+        }
+
         $user->syncRoles($data['role_name']);
 
         return $user;
@@ -74,7 +105,11 @@ class RoleRepository implements RoleInterface
 
     public function removeRoleFromUser($userId)
     {
-        $user = User::findOrFail($userId);
+        $user = User::find($userId);
+        if (! $user) {
+            return null;
+        }
+
         $user->syncRoles([]);
 
         return $user;
